l10n-brazil-master/requirements.txt
# generated from manifests external_dependencies
erpbrasil.assinatura
erpbrasil.base
erpbrasil.edoc
erpbrasil.edoc.pdf
erpbrasil.transmissao
nfelib
num2words
odoo_test_helper
pycep_correios
workalendar
xmldiff
cryptography
